feat: python ast loading in but broken
